requirements.txt
alembic==1.6.5
astroid==2.2.5
atomicwrites==1.4.1
attrs==23.1.0
awscli==1.27.118
botocore==1.29.118
cffi==1.15.1
click==8.0.1
colorama==0.4.4
cryptography==40.0.1
docutils==0.16
ecdsa==0.13.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
future==0.17.1
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==6.1.0
iniconfig==2.0.0
isort==4.3.18
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==1.0.1
lazy-object-proxy==1.4.0
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==23.1
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.3.0
PyYAML==5.4.1
rsa==4.7.2
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.18
toml==0.10.2
typed-ast==1.4.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.0.1
wrapt==1.11.1
zipp==3.15.0
python-dotenv==0.21.1